Pin the API to older versions of node-fetch and p-limit because they're ESM only and we're on CommonJS
Must use import to load ES Module: require() of ES modules is not supported

node-fetch is an ESM-only module - you are not able to import it with require. We recommend you stay on v2 which is built with CommonJS unless you use ESM yourself. We will continue to publish critical bug fixes for it.

https://stackoverflow.com/questions/69087292/requirenode-fetch-gives-err-require-esm